Separated BaseFormat into separate file
More docs on BaseFormat and the formats
class methods.
