+++ pydarn/io/borealis/borealis_formats.py
@@ -2,11 +2,10 @@
 # Author: Marci Detwiller
 """
 This file contains several classes with the fields that pertain to
-SuperDARN Borealis HDF5 files.
+SuperDARN Borealis HDF5 files. All formats inherit from BaseFormat.
 
 Classes
 -------
-BaseFormat
 BorealisRawacf
 BorealisBfiq
 BorealisAntennasIq
@@ -23,15 +22,6 @@ borealis_version_dict
     A lookup table for [version][filetype] that provides the appropriate class
     given the version and filetype strings.
 
-Design Concept
---------------
-All borealis format classes inherit from BaseFormat. All borealis format
-classes must contain specific methods that define the fields within
-that format. BaseFormat contains additional class methods that return useful
-information for restructuring all formats to array and site structure.
-The BaseFormat methods build off of the overwritten methods in each
-Borealis format class. See the BaseFormat docs for more information.
-
 Notes
 -----
 - Debug data files such as Borealis stage iq data (an intermediate
@@ -39,11 +29,6 @@ Notes
   progression from rawrf to output ptrs iq files) will not be included here.
   This is a debug format only and should not be used for higher level
   data products.
-- 'borealis_git_hash' and 'sqn_timestamps' are necessary fields for all
-  versions. 'borealis_git_hash' is necessary as its use is hardcoded into
-  the code in order to determine the format version to use. 'sqn_timestamps'
-  is necessary as all formats use this field to restructure from site to array
-  and vice versa.
 
 See Also
 --------
@@ -51,547 +36,31 @@ For more information on Borealis data files, see:
 https://borealis.readthedocs.io/en/latest/
 """
 
-from datetime import datetime
 import numpy as np
 
 from collections import OrderedDict
 
-
-class BaseFormat():
-    """
-    The base format of all Borealis format classes.
-
-    Class Methods
-    -------------
-    All Borealis formats should inherit from this class. The following class
-    methods should be specific to the format and therefore should be
-    overwritten by the format.
-
-    single_element_types(): dict
-        Dictionary of data field name to type for the format. This
-        dictionary should contain all fields in a record
-        that are not numpy arrays.
-    array_dtypes(): dict
-        Dictionary of data field names where numpy arrays are expected,
-        to numpy dtype for the format. This dictionary should contain all
-        fields that would be an array per record.
-    shared_fields(): list
-        List of the fields that are common (shared) across records. This
-        means that they can be reduced to a single value/array per file when
-        they are array restructured.
-    unshared_fields_dims_array(): dict
-        Unshared field: dimensions per record in array structure. Unshared
-        fields are not common across records. In array structure the first
-        dimension will be num_records followed by these dimensions. Dimensions
-        are provided as functions that will calculate the dimension given the
-        records or site data dictionary. This class method is used to convert
-        from site structure to array structure.
-    unshared_fields_dims_site(): dict
-        Unshared field: dimensions per record in site structure. Unshared
-        fields are not common across records. These dimensions can vary per
-        record so the functions take the arrays data dictionary and the
-        record number. This class method is used to convert from array
-        structure to site structure.
-    array_specific_fields_generate(): dict
-        Any fields that are array specific or require specific function to
-        generate. The key is the name of the array specific field and the
-        value in the dictionary is the function that takes the
-        records (site data dictionary) to generate the value for that field.
-        This class method is used when restructuring from site to array style.
-    site_specific_fields_generate(): dict
-        Any fields that are site specific or require specific function to
-        generate. The key is the name of the site specific field and the
-        value in the dictionary is the function that takes the arrays
-        (array data dictionary) and the record_num to generate the value for
-        that field at that record number. This class method is used when
-        restructuring from array to site style.
-
-    These following methods use the format-specific methods above to generate
-    their values and therefore should not be overwritten by the format class:
-
-    unshared_fields: list
-        List of the fields that are not common across records and therefore
-        must be stored as an array with first dimension = num_records in the
-        array structure.
-    array_specific_fields: list
-        List of fields that are only present in array files.
-    site_specific_fields: list
-        List of fields that are only present in site files.
-    site_fields: list
-        List of all fields that are in the site file type.
-    array_fields: list
-        List of all fields that are in the array file type.
-    site_single_element_fields : list
-        List of fields in the site files that are single element types.
-    site_single_element_types: dict
-        subset of single_element_types with only site keys.
-    site_array_dtypes_fields : list
-        List of fields in the site files that are made of numpy arrays.
-    site_array_dtypes: dict
-        subset of array_dtypes with only site keys.
-    array_single_element_fields : list
-        List of fields in the array files that are single element types.
-        Note that if the field is unshared it will appear as an
-        array of the type in the arrays data dictionary so no unshared fields
-        will appear in this list.
-    array_single_element_types: dict
-        subset of single_element_types with only array fields.
-    array_array_dtypes_fields : list
-        List of fields in the array files that are made of numpy arrays.
-        Includes fields that are single element but are unshared so are
-        converted to arrays in the array file.
-    array_array_dtypes: dict
-        fields in the array files that are made of numpy arrays, with their
-        given data type.
-    _site_to_array(data_dict): dict
-        Convert an OrderedDict of site data to array data using the information
-        provided for the specific data format.
-    _array_to_site(data_dict): OrderedDict
-        Convert a dictionary of array data to site data using the information
-        provided for the specific data format.
-
-    Static Methods
-    --------------
-    find_max_sequences(records): int
-        Find the max number of sequences between records in a site file, for
-        restructuring to arrays.
-    find_max_beams(records): int
-        Find the max number of beams between records in a site file, for
-        restructuring to arrays.
-    find_max_blanked_samples(records): int
-        Find the max number of blanked samples between records in a site file,
-        for restructuring to arrays.
-    reshape_site_arrays(records): OrderedDict
-        Returns the site data with the arrays reshaped. Some site data arrays
-        may be stored in linear dimensions, so this reshapes any if needed.
-        Should be overwritten by the child class.
-
-    Notes
-    -----
-    single_element_types.keys() + array_dtypes.keys() = all known fields
-    shared_fields + unshared_fields + array_only_fields = all fields in
-        array file
-    shared_fields + unshared_fields + site_only_fields = all fields in
-        site file
-    """
-
-    @staticmethod
-    def find_max_sequences(records: OrderedDict) -> int:
-        """
-        Finds the maximum number of sequences between records in a Borealis
-        site style records file.
-
-        Parameters
-        ----------
-        records
-            Site formatted records from a Borealis file, organized as one
-            record for each slice
-
-        Returns
-        -------
-        max_seqs
-            The largest number of sequences found in one record from the
-            file
-        """
-        max_seqs = 0
-        for k in records:
-            if max_seqs < records[k]["num_sequences"]:
-                max_seqs = records[k]["num_sequences"]
-        return max_seqs
-
-    @staticmethod
-    def find_max_beams(records: OrderedDict) -> int:
-        """
-        Finds the maximum number of beams between records in a Borealis
-        site style records file.
-
-        Parameters
-        ----------
-        records
-            Site formatted records from a Borealis file, organized as one
-            record for each slice
-
-        Returns
-        -------
-        max_beams
-            The largest number of beams found in one record from the
-            file
-        """
-        max_beams = 0
-        for k in records:
-            if max_beams < len(records[k]["beam_nums"]):
-                max_beams = len(records[k]["beam_nums"])
-        return max_beams
-
-    @staticmethod
-    def find_max_blanked_samples(records: OrderedDict) -> int:
-        """
-        Finds the maximum number of blanked samples between records in a
-        Borealis site style records file.
-
-        Parameters
-        ----------
-        records
-            Site formatted records from a Borealis file, organized as one
-            record for each slice
-
-        Returns
-        -------
-        max_beams
-            The largest number of beams found in one record from the
-            file
-        """
-        max_blanked_samples = 0
-        for k in records:
-            if max_blanked_samples < len(records[k]["blanked_samples"]):
-                max_blanked_samples = len(records[k]["blanked_samples"])
-        return max_blanked_samples
-
-    @staticmethod
-    def reshape_site_arrays(records: OrderedDict) -> OrderedDict:
-        """
-        This is a necessary function for interpreting site data, as
-        some arrays in site data have been flattened in the file and
-        need to be reshaped to be interpreted correctly.
-
-        This reshapes them to the correct dimensions. Some formats may not
-        have this issue, in which case this function does not need to be
-        updated by the child class.
-
-        This function is used in the _site_to_array restructuring which
-        is common to all formats.
-
-        Parameters
-        ----------
-        records
-            site-style records dictionary.
-
-        Returns
-        -------
-        records
-            site-style records dictionary with any data field arrays reshaped
-            as required by the format. The BaseFormat does not contain
-            any fields that need to be reshaped but some child classes do.
-        """
-        return records
-
-    @staticmethod
-    def flatten_site_arrays(records: OrderedDict) -> OrderedDict:
-        """
-        This is a necessary function for writing site data, as
-        some arrays in site data are flattened in the file and
-        have been reshaped to be interpreted correctly.
-
-        This flattens the data correctly for the file. Some formats may not
-        have this issue, in which case this function does not need to be
-        updated by the child class.
-
-        This function is used in the _array_to_site restructuring which
-        is common to all formats.
-
-        Parameters
-        ----------
-        records
-            site-style records dictionary.
-
-        Returns
-        -------
-        records
-            site-style records dictionary with any data field arrays flattened
-            as required by the format to be written to file. The BaseFormat
-            does not contain any fields that need to be flattened for writing
-            but some child classes do.
-        """
-        return records
-
-    # initialize these methods, which should be rewritten by the format.
-    @classmethod
-    def single_element_types(cls):
-        return {
-            # Identifies the version of Borealis that made this data.
-            # Necessary for all versions and formats.
-            "borealis_git_hash": np.unicode_,
-            }
-
-    @classmethod
-    def array_dtypes(cls):
-        """
-        The following fields are necessary for ALL FORMATS.
-        """
-        return {
-            # A list of GPS timestamps of the beginning of transmission for
-            # each sampling period in the integration time. Seconds since
-            # epoch. Necessary for all formats.
-            "sqn_timestamps": np.float64,
-            }
-
-    @classmethod
-    def shared_fields(cls):
-        return []
-
-    @classmethod
-    def unshared_fields_dims_array(cls):
-        return {}
-
-    @classmethod
-    def unshared_fields_dims_site(cls):
-        return {}
-
-    @classmethod
-    def array_specific_fields_generate(cls):
-        return {}
-
-    @classmethod
-    def site_specific_fields_generate(cls):
-        return {}
-
-    # these methods build off of the above.
-    @classmethod
-    def unshared_fields(cls):
-        return list(cls.unshared_fields_dims_array().keys())
-
-    @classmethod
-    def array_specific_fields(cls):
-        return list(cls.array_specific_fields_generate().keys())
-
-    @classmethod
-    def site_specific_fields(cls):
-        return list(cls.site_specific_fields_generate().keys())
-
-    @classmethod
-    def site_fields(cls):
-        """ All site fields """
-        return cls.shared_fields() + cls.unshared_fields() + \
-            cls.site_specific_fields()
-
-    @classmethod
-    def array_fields(cls):
-        """ All array fields """
-        return cls.shared_fields() + cls.unshared_fields() + \
-            cls.array_specific_fields()
-
-    @classmethod
-    def site_single_element_fields(cls):
-        """ All site fields that are single element in a list """
-        return [k for k in cls.site_fields() if k in
-                list(cls.single_element_types().keys())]
-
-    @classmethod
-    def site_single_element_types(cls):
-        """ Dict of site single element field: type"""
-        return {k: cls.single_element_types()[k]
-                for k in cls.site_single_element_fields()}
-
-    @classmethod
-    def site_array_dtypes_fields(cls):
-        """ All site fields that are arrays in a list """
-        return [k for k in cls.site_fields() if k in
-                list(cls.array_dtypes().keys())]
-
-    @classmethod
-    def site_array_dtypes(cls):
-        """ Dict of site array field : dtype """
-        return {k: cls.array_dtypes()[k] for k in
-                cls.site_array_dtypes_fields()}
-
-    # for single element fields in the array filetypes, they must
-    # be a shared field.
-    @classmethod
-    def array_single_element_fields(cls):
-        """ List of array restructured single element fields """
-        return [k for k in cls.array_fields() if
-                k in list(cls.single_element_types().keys()) and k in
-                cls.shared_fields()]
-
-    @classmethod
-    def array_single_element_types(cls):
-        """ Dict of array restructured single element field : type """
-        return {k: cls.single_element_types()[k]
-                for k in cls.array_single_element_fields()}
-
-    # for array filetypes, there are more array dtypes for any unshared
-    # fields. If the field was a single_element_type and is unshared,
-    # it is now an array of num_records length.
-    @classmethod
-    def array_array_dtypes_fields(cls):
-        """ List of array restructured array fields """
-        return [k for k in cls.array_fields() if
-                k in list(cls.array_dtypes().keys())] + \
-               [k for k in cls.array_fields() if k in
-                list(cls.single_element_types().keys()) and
-                ((k in cls.unshared_fields()) or
-                 (k in cls.array_specific_fields()))]
-
-    @classmethod
-    def array_array_dtypes(cls):
-        """ Dict of array restructured array field : dtype """
-        array_array_dtypes = {k: cls.array_dtypes()[k] for k in
-                              cls.array_array_dtypes_fields() if k in
-                              list(cls.array_dtypes().keys())}
-
-        array_array_dtypes.update(
-            {k: cls.single_element_types()[k] for
-             k in cls.array_array_dtypes_fields() if k in
-             list(cls.single_element_types().keys())})
-
-        return array_array_dtypes
-
-    @classmethod
-    def _site_to_array(cls, data_dict: OrderedDict) -> dict:
-        """
-        Base function for converting site Borealis data to
-        restructured array format.
-
-        Parameters
-        ----------
-        data_dict: OrderedDict
-            a dict of timestamped records loaded from an hdf5 Borealis site
-            file
-
-        Returns
-        -------
-        new_data_dict
-            A dictionary containing the data from data_dict
-            reformatted to be stored entirely in array style, or as
-            one entry if the field does not change between records.
-            This means that for fields that change between records,
-            the first dimension in the array will equal num_records
-            (these are called unshared_fields). For fields common to all
-            records, there will only be the one value that applies (these
-            are known as shared_fields).
-        """
-        new_data_dict = dict()
-        num_records = len(data_dict)
-
-        # some fields are linear in site style and need to be reshaped.
-        data_dict = cls.reshape_site_arrays(data_dict)
-
-        # write shared fields to dictionary
-        first_key = list(data_dict.keys())[0]
-        for field in cls.shared_fields():
-            new_data_dict[field] = data_dict[first_key][field]
-
-        # write array specific fields using the given functions.
-        for field in cls.array_specific_fields():
-            new_data_dict[field] = cls.array_specific_fields_generate(
-                )[field](data_dict)
-
-        # write the unshared fields, initializing empty arrays to start.
-        temp_array_dict = dict()
-
-        # get array dims of the unshared fields arrays
-        field_dimensions = {}
-        for field in cls.unshared_fields():
-            dims = [dimension_function(data_dict) for
-                    dimension_function in
-                    cls.unshared_fields_dims_array()[field]]
-            field_dimensions[field] = dims
-
-        # all fields to become arrays
-        for field, dims in field_dimensions.items():
-            array_dims = [num_records] + dims
-            array_dims = tuple(array_dims)
-
-            if field in cls.single_element_types():
-                datatype = cls.single_element_types()[field]
-            else:  # field in array_dtypes
-                datatype = cls.array_dtypes()[field]
-            empty_array = np.empty(array_dims, dtype=datatype)
-            # initialize all values to NaN; some indices may not be filled
-            # do to dimensions that are max values (num sequences, etc can
-            # change between records)
-            empty_array[:] = np.NaN
-            temp_array_dict[field] = empty_array
-
-        # iterate through the records, filling the unshared and array only
-        # fields
-        for rec_idx, k in enumerate(data_dict.keys()):
-            for field in cls.unshared_fields():  # all unshared fields
-                empty_array = temp_array_dict[field]
-                if type(data_dict[first_key][field]) == np.ndarray:
-                    # only fill the correct length, appended NaNs occur for
-                    # dims with a determined max value
-                    data_buffer = data_dict[k][field]
-                    buffer_shape = data_buffer.shape
-                    index_slice = [slice(0, i) for i in buffer_shape]
-                    # insert record index at start of array's slice list
-                    index_slice.insert(0, rec_idx)
-                    index_slice = tuple(index_slice)
-                    # place data buffer in the correct place
-                    empty_array[index_slice] = data_buffer
-                else:  # not an array, num_records is the only dimension
-                    empty_array[rec_idx] = data_dict[k][field]
-
-        new_data_dict.update(temp_array_dict)
-
-        return new_data_dict
-
-    @classmethod
-    def _array_to_site(cls, data_dict: dict) -> OrderedDict:
-        """
-        Base function for converting array Borealis data to
-        site format.
-
-        Parameters
-        ----------
-        data_dict: dictionary of array restructured Borealis data.
-
-        Returns
-        -------
-        new_data_dict
-            An OrderedDict of timestamped records as if loaded from
-            the original site file.
-        """
-        timestamp_dict = OrderedDict()
-        for record_num, seq_timestamp in \
-                enumerate(data_dict["sqn_timestamps"]):
-            # format dictionary key in the same way it is done
-            # in datawrite on site
-            seq_datetime = datetime.utcfromtimestamp(seq_timestamp[0])
-            epoch = datetime.utcfromtimestamp(0)
-            key = str(int((seq_datetime - epoch).total_seconds() * 1000))
-
-            timestamp_dict[key] = dict()
-            # populate shared fields in each record,
-            for field in cls.shared_fields():
-                timestamp_dict[key][field] = data_dict[field]
-
-            # populate site specific fields using given functions
-            # that take both the arrays data and the record number
-            for field in cls.site_specific_fields():
-                timestamp_dict[key][field] = cls.site_specific_fields_generate(
-                    )[field](data_dict, record_num)
-
-            for field in cls.unshared_fields():
-                if field in cls.single_element_types():
-                    datatype = cls.single_element_types()[field]
-                    # field is not an array, single element per record.
-                    # unshared_field_dims_site should give empty list.
-                    timestamp_dict[key][field] = datatype(data_dict[field]
-                                                          [record_num])
-                else:  # field in array_dtypes
-                    datatype = cls.array_dtypes()[field]
-                    # need to get the dims correct, not always equal to the max
-                    site_dims = [dimension_function(data_dict, record_num)
-                                 for dimension_function in
-                                 cls.unshared_fields_dims_site()[field]]
-                    index_slice = [slice(0, i) for i in site_dims]
-                    index_slice.insert(0, record_num)
-                    index_slice = tuple(index_slice)
-                    timestamp_dict[key][field] = data_dict[field][index_slice]
-
-        timestamp_dict = cls.flatten_site_arrays(timestamp_dict)
-
-        return timestamp_dict
+from .base_format import BaseFormat
 
 
 class BorealisRawacfv0_4(BaseFormat):
     """
     Class containing Borealis Rawacf data fields and their types.
 
+    See Also
+    --------
+    BaseFormat
+    BorealisRawacf (most up to date format)
+
+    Notes
+    -----
     Rawacf data has been mixed, filtered, and decimated; beamformed and
     combined into antenna arrays; then autocorrelated and correlated between
     antenna arrays to produce matrices of num_ranges x num_lags.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
 
     Static Methods
     --------------
@@ -668,7 +137,15 @@ class BorealisRawacfv0_4(BaseFormat):
 
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+        """
         return {
             # Identifies the version of Borealis that made this data. Necessary
             # for all versions.
@@ -722,7 +199,15 @@ class BorealisRawacfv0_4(BaseFormat):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+        """
         return {
             # The pulse sequence in multiples of the tau_spacing.
             "pulses": np.uint32,
@@ -853,10 +338,21 @@ class BorealisBfiqv0_4(BaseFormat):
     """
     Class containing Borealis Bfiq data fields and their types.
 
+    See Also
+    --------
+    BaseFormat
+    BorealisBfiq (most up to date format)
+
+    Notes
+    -----
     Bfiq data is beamformed i and q data. It has been mixed, filtered,
     decimated to the final output receive rate, and it has been beamformed
     and all channels have been combined into their arrays. No correlation
     or averaging has occurred.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
 
     Static Methods
     --------------
@@ -926,7 +422,15 @@ class BorealisBfiqv0_4(BaseFormat):
 
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+        """
         return {
             # Identifies the version of Borealis that made this data. Necessary
             # for all versions.
@@ -984,7 +488,15 @@ class BorealisBfiqv0_4(BaseFormat):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+        """
         return {
             # The pulse sequence in multiples of the tau_spacing.
             "pulses": np.uint32,
@@ -1105,9 +617,20 @@ class BorealisAntennasIqv0_4(BaseFormat):
     """
     Class containing Borealis Antennas iq data fields and their types.
 
+    See Also
+    --------
+    BaseFormat
+    BorealisAntennasIq (most up to date format)
+
+    Notes
+    -----
     Antennas iq data is data with all channels separated. It has been mixed
     and filtered, but it has not been beamformed or combined into the
     entire antenna array data product.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
 
     Static Methods
     --------------
@@ -1164,10 +687,32 @@ class BorealisAntennasIqv0_4(BaseFormat):
     @staticmethod
     def flatten_site_arrays(records: OrderedDict) -> OrderedDict:
         """
-        Sometimes arrays in the site style have been reduced to
-        linear arrays with given dimensions. This function returns
-        the site data with all arrays flattened as desired for
-        storing in the site format.
+        A function to flatten record fields if needed for storing in the
+        site structure. The fields for flattening are specific to the
+        format.
+
+        Parameters
+        ----------
+        records
+            An OrderedDict of the site style data, organized
+            by record. Records are stored with timestamps
+            as the keys and the data for that timestamp
+            stored as a dictionary.
+
+        Returns
+        -------
+        records
+            An OrderedDict of the site style data, with the data
+            field in all records flattened as is the convention
+            in site structured files.
+
+        Notes
+        -----
+        This function is used by the _arrays_to_site class method
+        in restructuring files.
+
+        BorealisAntennasIq has the data field flattened in the
+        site structured files.
         """
         for key in list(records.keys()):
             for field in ['data']:
@@ -1177,7 +722,15 @@ class BorealisAntennasIqv0_4(BaseFormat):
 
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+        """
         return {
             # Identifies the version of Borealis that made this data. Necessary
             # for all versions.
@@ -1227,7 +780,15 @@ class BorealisAntennasIqv0_4(BaseFormat):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+        """
         return {
             # The pulse sequence in multiples of the tau_spacing.
             "pulses": np.uint32,
@@ -1336,30 +897,71 @@ class BorealisRawrfv0_4(BaseFormat):
     """
     Class containing Borealis Rawrf data fields and their types.
 
+    See Also
+    --------
+    BaseFormat
+    BorealisRawrf (most up to date format)
+
+    Notes
+    -----
     This filetype only has site files and cannot be restructured.
 
     Rawrf data is data that has been produced at the original receive bandwidth
     and has not been mixed, filtered, or decimated.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
 
     Static Methods
     --------------
     is_restructureable: bool
         Returns False, this format cannot be restructured.
     reshape_site_arrays(OrderedDict): OrderedDict
-        Reshapes the data field according to data dimensions.
+        Reshapes the data field in the records according to data
+        dimensions.
+    flatten_site_arrays(OrderedDict): OrderedDict
+        Flattens the data field in the records as is the
+        convention for site style files.
     """
 
     @staticmethod
     def is_restructureable() -> bool:
+        """
+        Returns if the format can be restructured from site to array
+        structure. Some file formats have not been designed to be
+        restructured to the array structure.
+
+        Notes
+        -----
+        BorealisRawrf is a very uncommon format and therefore has
+        not been implemented to be converted to arrays.
+        """
         return False
 
     @staticmethod
     def reshape_site_arrays(records: OrderedDict) -> OrderedDict:
         """
-        Sometimes arrays in the site style have been reduced to
-        linear arrays with given dimensions. This function returns
-        the site data with all arrays reshaped correctly as
-        per the format.
+        See BaseFormat class for description and use of this method.
+
+        Parameters
+        ----------
+        records
+            An OrderedDict of the site style data, organized
+            by record. Records are stored with timestamps
+            as the keys and the data for that timestamp
+            stored as a dictionary.
+
+        Returns
+        -------
+        records
+            An OrderedDict of the site style data, with the data
+            field in all records reshaped to the correct dimensions.
+
+        Notes
+        -----
+        BorealisRawrf has the data field flattened in the
+        site structured files, so this field is reshaped in here.
         """
         for key in list(records.keys()):
             record_dimensions = records[key]['data_dimensions']
@@ -1372,10 +974,27 @@ class BorealisRawrfv0_4(BaseFormat):
     @staticmethod
     def flatten_site_arrays(records: OrderedDict) -> OrderedDict:
         """
-        Sometimes arrays in the site style have been reduced to
-        linear arrays with given dimensions. This function returns
-        the site data with all arrays flattened as desired for
-        storing in the site format.
+        See BaseFormat class for description and use of this method.
+
+        Parameters
+        ----------
+        records
+            An OrderedDict of the site style data, organized
+            by record. Records are stored with timestamps
+            as the keys and the data for that timestamp
+            stored as a dictionary.
+
+        Returns
+        -------
+        records
+            An OrderedDict of the site style data, with the data
+            field in all records flattened as is the convention
+            in site structured files.
+
+        Notes
+        -----
+        BorealisRawrf has the data field flattened in the
+        site structured files.
         """
         for key in list(records.keys()):
             for field in ['data']:
@@ -1385,7 +1004,15 @@ class BorealisRawrfv0_4(BaseFormat):
 
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+        """
         return {
             # Identifies the version of Borealis that made this data. Necessary
             # for all versions.
@@ -1423,7 +1050,15 @@ class BorealisRawrfv0_4(BaseFormat):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+        """
         return {
             # A list of GPS timestamps of the beginning of transmission for
             # each sampling period in the integration time. Seconds since
@@ -1460,7 +1095,7 @@ class BorealisRawacf(BorealisRawacfv0_4):
 
     See BaseFormat for description of classmethods and how they
     are used to verify format files and restructure Borealis files to
-    array and site formats.
+    array and site structure.
 
     In v0.5, the following fields were added:
     slice_id, slice_interfacing, scheduling_mode, and averaging_method.
@@ -1468,11 +1103,25 @@ class BorealisRawacf(BorealisRawacfv0_4):
     restructuring.
     """
 
-    # New fields added in v0.5 to shared fields:
-
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+
+        Notes
+        -----
+        In Borealis v0.5, slice_id, slice_interfacing, scheduling_mode, and
+        averaging_methods were added as single element fields.
+        num_blanked_samples is an array-specific field as the number may vary
+        from record to record, but it is also a single-element-per-record
+        field.
+        """
         single_element_types = super(BorealisRawacf,
                                      cls).single_element_types()
         single_element_types.update({
@@ -1539,12 +1188,21 @@ class BorealisBfiq(BorealisBfiqv0_4):
     Class containing Borealis Bfiq data fields and their types for the
     current version of Borealis (v0.5).
 
+    See Also
+    --------
+    BaseFormat
+    BorealisBfiqv0_4
+
+    Notes
+    -----
     Bfiq data is beamformed i and q data. It has been mixed, filtered,
     decimated to the final output receive rate, and it has been beamformed
     and all channels have been combined into their arrays. No correlation
     or averaging has occurred.
 
-    See base class for description of attributes and notes.
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
 
     In v0.5, the following fields were added:
     slice_id, slice_interfacing, and scheduling_mode.
@@ -1552,11 +1210,24 @@ class BorealisBfiq(BorealisBfiqv0_4):
     restructuring.
     """
 
-    # New fields added in v0.5 to shared fields:
-
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+
+        Notes
+        -----
+        In Borealis v0.5, slice_id, slice_interfacing, and scheduling_mode
+        were added as single element fields. num_blanked_samples is an
+        array-specific field as the number may vary from record to record,
+        but it is also a single-element-per-record field.
+        """
         single_element_types = super(BorealisBfiq,
                                      cls).single_element_types()
         single_element_types.update({
@@ -1621,20 +1292,44 @@ class BorealisAntennasIq(BorealisAntennasIqv0_4):
     Class containing Borealis Antennas iq data fields and their types for
     Borealis current version (v0.5).
 
+    See Also
+    --------
+    BaseFormat
+    BorealisAntennasIqv0_4
+
+    Notes
+    -----
     Antennas iq data is data with all channels separated. It has been mixed
     and filtered, but it has not been beamformed or combined into the
     entire antenna array data product.
 
-    See base class for description of attributes and notes.
-    In v0.5, the following fields were added:
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
+    In v0.5, the following fields were added to the Borealis-produced
+    site structured files:
     slice_id, slice_interfacing, scheduling_mode, and blanked_samples.
     """
 
-    # New fields added in v0.5 to shared fields:
-
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+
+        Notes
+        -----
+        In Borealis v0.5, slice_id, slice_interfacing, and scheduling_mode
+        were added as single element fields. num_blanked_samples is an
+        array-specific field as the number may vary from record to record,
+        but it is also a single-element-per-record field.
+        """
         single_element_types = super(BorealisAntennasIq,
                                      cls).single_element_types()
         single_element_types.update({
@@ -1652,7 +1347,20 @@ class BorealisAntennasIq(BorealisAntennasIqv0_4):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+
+        Notes
+        -----
+        In Borealis v0.5, blanked_samples was added as an array
+        field.
+        """
         array_dtypes = super(BorealisAntennasIq, cls).array_dtypes()
         array_dtypes.update({
             # Samples that occur during TR switching (transmission times)
@@ -1708,16 +1416,37 @@ class BorealisRawrf(BorealisRawrfv0_4):
     Class containing Borealis Rawrf data fields and their types for current
     Borealis version (v0.5).
 
-    See base class for description of attributes and notes.
+    See Also
+    --------
+    BaseFormat
+    BorealisRawrfv0_4
+
+    Notes
+    -----
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
     In v0.5, the following fields were added:
     slice_id, slice_interfacing, scheduling_mode, and blanked_samples.
     """
 
-    # New fields added in v0.5 to shared fields:
-
     @classmethod
     def single_element_types(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        single_element_types
+            All the single-element fields in records of the
+            format, as a dictionary fieldname : type.
+
+        Notes
+        -----
+        In Borealis v0.5, scheduling_mode was added as a single
+        element field.
+        """
         single_element_types = super(BorealisRawrf,
                                      cls).single_element_types()
         single_element_types.update({
@@ -1729,7 +1458,20 @@ class BorealisRawrf(BorealisRawrfv0_4):
 
     @classmethod
     def array_dtypes(cls):
-        """ See BaseFormat class for description. """
+        """
+        See BaseFormat class for description and use of this method.
+
+        Returns
+        -------
+        array_dtypes
+            All the array fields in records of the
+            format, as a dictionary fieldname : array dtype.
+
+        Notes
+        -----
+        In Borealis v0.5, blanked_samples was added as an array
+        field.
+        """
         array_dtypes = super(BorealisRawrf, cls).array_dtypes()
         array_dtypes.update({
             # Samples that occur during TR switching (transmission times)
